coll/HAN: Add support for XHC's "special" Reduce for the low-comm in Allreduce
MPI_Reduce in XHC is not complete; it is implemented as a sub-case of
Allreduce, and requires that the rbuf parameter is always present and
appropriately sized for all ranks (not only for the root). This
implementation is disabled by default and falls back to another
coll component, but can be manually enabled for a single operation
by invoking it with root=-1, which will do a reduce to rank 0.

Inside HAN's Allreduce, the rbuf parameter restriction is satisfied,
so it's safe to use this partially implemented Reduce.

This patch is temporary (TM) until XHC's Reduce is fully implemented.
The reason for its existence is the improved Allreduce performance
potential with XHC for the intra-comm.

Signed-off-by: George Katevenis <gkatev@ics.forth.gr>